+++ pointercrate-user/src/auth/mod.rs
@@ -6,14 +6,10 @@
 //! * Modification of own account
 
 pub use self::patch::PatchMe;
-#[cfg(feature = "legacy_accounts")]
-pub use self::post::Registration;
-use crate::{
-    error::{Result, UserError},
-    User,
-};
+use crate::{error::Result, User};
 use jsonwebtoken::{DecodingKey, EncodingKey};
-use log::{debug, warn};
+use legacy::LegacyAuthenticatedUser;
+use log::warn;
 use pointercrate_core::error::CoreError;
 use serde::{Deserialize, Serialize};
 use std::{
@@ -23,16 +19,11 @@ use std::{
 
 mod delete;
 mod get;
+pub mod legacy;
 mod patch;
-mod post;
 
-pub struct AuthenticatedUser {
-    user: User,
-    auth_method: AuthenticationMethod,
-}
-
-pub enum AuthenticationMethod {
-    Legacy { password_hash: String },
+pub enum AuthenticatedUser {
+    Legacy(LegacyAuthenticatedUser),
 }
 
 #[derive(Debug, Deserialize, Serialize, Copy, Clone)]
@@ -49,31 +40,27 @@ pub struct CSRFClaims {
 
 impl AuthenticatedUser {
     pub fn into_inner(self) -> User {
-        self.user
+        match self {
+            AuthenticatedUser::Legacy(legacy) => legacy.into_user(),
+        }
     }
 
     pub fn inner(&self) -> &User {
-        &self.user
-    }
-
-    pub fn validate_password(password: &str) -> Result<()> {
-        if password.len() < 10 {
-            return Err(UserError::InvalidPassword);
+        match self {
+            AuthenticatedUser::Legacy(legacy) => legacy.user(),
         }
-
-        Ok(())
     }
 
     fn jwt_secret(&self) -> Vec<u8> {
         let mut key: Vec<u8> = pointercrate_core::config::secret();
-        key.extend(self.password_salt());
+        key.extend(self.salt());
         key
     }
 
     pub fn generate_access_token(&self) -> String {
         jsonwebtoken::encode(
             &jsonwebtoken::Header::default(),
-            &AccessClaims { id: self.user.id },
+            &AccessClaims { id: self.inner().id },
             &EncodingKey::from_secret(&self.jwt_secret()),
         )
         .unwrap()
@@ -87,13 +74,13 @@ impl AuthenticatedUser {
 
         jsonwebtoken::decode::<AccessClaims>(token, &DecodingKey::from_secret(&self.jwt_secret()), &validation)
             .map_err(|err| {
-                warn!("Token validation FAILED for account {}: {}", self.user, err);
+                warn!("Token validation FAILED for account {}: {}", self.inner(), err);
 
                 CoreError::Unauthorized.into()
             })
             .and_then(move |token_data| {
                 // sanity check, should never fail
-                if token_data.claims.id != self.user.id {
+                if token_data.claims.id != self.inner().id {
                     log::error!(
                         "Access token for user {} decoded successfully even though user {} is logged in",
                         token_data.claims.id,
@@ -114,7 +101,7 @@ impl AuthenticatedUser {
             .expect("time went backwards (and this is probably gonna bite me in the ass when it comes to daytimesaving crap)");
 
         let claim = CSRFClaims {
-            id: self.user.id,
+            id: self.inner().id,
             iat: since_epoch.as_secs(),
             exp: (since_epoch + Duration::from_secs(3600)).as_secs(),
         };
@@ -134,15 +121,16 @@ impl AuthenticatedUser {
 
         jsonwebtoken::decode::<CSRFClaims>(token, &DecodingKey::from_secret(&pointercrate_core::config::secret()), &validation)
             .map_err(|err| {
-                warn!("Access token validation FAILED for account {}: {}", self.user, err);
+                warn!("Access token validation FAILED for account {}: {}", self.inner(), err);
 
                 CoreError::Unauthorized.into()
             })
             .and_then(|token_data| {
-                if token_data.claims.id != self.user.id {
+                if token_data.claims.id != self.inner().id {
                     warn!(
                         "User {} attempt to authenticate using CSRF token generated for user {}",
-                        self.user, token_data.claims.id
+                        self.inner(),
+                        token_data.claims.id
                     );
 
                     Err(CoreError::Unauthorized.into())
@@ -152,41 +140,15 @@ impl AuthenticatedUser {
             })
     }
 
-    fn password_salt(&self) -> Vec<u8> {
-        match &self.auth_method {
-            AuthenticationMethod::Legacy { password_hash } => {
-                let raw_parts: Vec<_> = password_hash.split('$').filter(|s| !s.is_empty()).collect();
-
-                match &raw_parts[..] {
-                    [_, _, hash] => b64::decode(&hash[..22]),
-                    _ => unreachable!(),
-                }
-            },
-            _ => Vec::new(),
+    fn salt(&self) -> Vec<u8> {
+        match self {
+            AuthenticatedUser::Legacy(legacy) => legacy.salt(),
         }
     }
 
     pub fn verify_password(self, password: &str) -> Result<Self> {
-        match &self.auth_method {
-            AuthenticationMethod::Legacy { password_hash } => {
-                debug!("Verifying a password!");
-
-                let valid = bcrypt::verify(password, password_hash).map_err(|err| {
-                    warn!("Password verification FAILED for account {}: {}", self.user, err);
-
-                    UserError::Core(CoreError::Unauthorized)
-                })?;
-
-                if valid {
-                    debug!("Password correct, proceeding");
-
-                    Ok(self)
-                } else {
-                    warn!("Potentially malicious log-in attempt to account {}", self.user);
-
-                    Err(CoreError::Unauthorized.into())
-                }
-            },
+        match &self {
+            AuthenticatedUser::Legacy(legacy) => legacy.verify(password).map(|_| self),
             _ => Err(CoreError::Unauthorized.into()),
         }
     }
@@ -196,36 +158,30 @@ impl AuthenticatedUser {
 mod tests {
     use crate::{AuthenticatedUser, User};
 
-    use super::AuthenticationMethod;
-
     fn patrick() -> AuthenticatedUser {
-        AuthenticatedUser {
-            user: User {
+        AuthenticatedUser::legacy(
+            User {
                 id: 0,
                 name: "Patrick".to_string(),
                 permissions: 0,
                 display_name: None,
                 youtube_channel: None,
             },
-            auth_method: AuthenticationMethod::Legacy {
-                password_hash: bcrypt::hash("bad password", bcrypt::DEFAULT_COST).unwrap(),
-            },
-        }
+            bcrypt::hash("bad password", bcrypt::DEFAULT_COST).unwrap(),
+        )
     }
 
     fn jacob() -> AuthenticatedUser {
-        AuthenticatedUser {
-            user: User {
+        AuthenticatedUser::legacy(
+            User {
                 id: 1,
-                name: "Jacob".to_string(),
+                name: "".to_string(),
                 permissions: 0,
                 display_name: None,
                 youtube_channel: None,
             },
-            auth_method: AuthenticationMethod::Legacy {
-                password_hash: bcrypt::hash("bad password", bcrypt::DEFAULT_COST).unwrap(),
-            },
-        }
+            bcrypt::hash("bad password", bcrypt::DEFAULT_COST).unwrap(),
+        )
     }
 
     #[test]
@@ -259,109 +215,5 @@ mod tests {
 
         assert!(patrick.validate_access_token(&patricks_access_token).is_ok());
         assert!(jacob.validate_access_token(&patricks_access_token).is_err());
-    }
-}
-
-// This code is copied from https://github.com/Keats/rust-bcrypt/blob/master/src/b64.rs
-// with slight modifications (removal of `encode` and error handling)
-mod b64 {
-    use std::collections::HashMap;
-
-    use base64::{engine::general_purpose::STANDARD, Engine};
-    use lazy_static::lazy_static;
-
-    // Decoding table from bcrypt base64 to standard base64 and standard -> bcrypt
-    // Bcrypt has its own base64 alphabet
-    // ./ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789
-    lazy_static! {
-        #[allow(unused_results)]
-        static ref BCRYPT_TO_STANDARD: HashMap<char, &'static str> = {
-            let mut m = HashMap::new();
-            m.insert('/', "B");
-            m.insert('.', "A");
-            m.insert('1', "3");
-            m.insert('0', "2");
-            m.insert('3', "5");
-            m.insert('2', "4");
-            m.insert('5', "7");
-            m.insert('4', "6");
-            m.insert('7', "9");
-            m.insert('6', "8");
-            m.insert('9', "/");
-            m.insert('8', "+");
-            m.insert('A', "C");
-            m.insert('C', "E");
-            m.insert('B', "D");
-            m.insert('E', "G");
-            m.insert('D', "F");
-            m.insert('G', "I");
-            m.insert('F', "H");
-            m.insert('I', "K");
-            m.insert('H', "J");
-            m.insert('K', "M");
-            m.insert('J', "L");
-            m.insert('M', "O");
-            m.insert('L', "N");
-            m.insert('O', "Q");
-            m.insert('N', "P");
-            m.insert('Q', "S");
-            m.insert('P', "R");
-            m.insert('S', "U");
-            m.insert('R', "T");
-            m.insert('U', "W");
-            m.insert('T', "V");
-            m.insert('W', "Y");
-            m.insert('V', "X");
-            m.insert('Y', "a");
-            m.insert('X', "Z");
-            m.insert('Z', "b");
-            m.insert('a', "c");
-            m.insert('c', "e");
-            m.insert('b', "d");
-            m.insert('e', "g");
-            m.insert('d', "f");
-            m.insert('g', "i");
-            m.insert('f', "h");
-            m.insert('i', "k");
-            m.insert('h', "j");
-            m.insert('k', "m");
-            m.insert('j', "l");
-            m.insert('m', "o");
-            m.insert('l', "n");
-            m.insert('o', "q");
-            m.insert('n', "p");
-            m.insert('q', "s");
-            m.insert('p', "r");
-            m.insert('s', "u");
-            m.insert('r', "t");
-            m.insert('u', "w");
-            m.insert('t', "v");
-            m.insert('w', "y");
-            m.insert('v', "x");
-            m.insert('y', "0");
-            m.insert('x', "z");
-            m.insert('z', "1");
-            m
-        };
-    }
-
-    // Can potentially panic if the hash given contains invalid characters
-    pub(super) fn decode(hash: &str) -> Vec<u8> {
-        let mut res = String::with_capacity(hash.len());
-        for ch in hash.chars() {
-            res.push_str(BCRYPT_TO_STANDARD.get(&ch).unwrap())
-        }
-
-        // Bcrypt base64 has no padding but standard has
-        // so we need to actually add padding ourselves
-        if hash.len() % 4 > 0 {
-            let padding = 4 - hash.len() % 4;
-            for _ in 0..padding {
-                res.push('=');
-            }
-        }
-
-        // if we had non standard chars, it would have errored before
-        STANDARD.decode(&res).unwrap()
     }
 }
